requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.6.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.2
promise==2.3
PyJWT==1.7.1
pytz==2021.1
ratelimit==2.2.1
requests==2.26.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.3.0
